Добавлен алгоритм следования вдоль периметра

+++ robot/utils/map.cpp
@@ -1,7 +1,7 @@
 #include "map.h"
 
-uint16_t map(uint16_t src, uint8_t src_min, uint8_t src_max,
-        uint8_t dst_min, uint8_t dst_max)
+uint16_t map(uint16_t src, uint16_t src_min, uint16_t src_max,
+        uint16_t dst_min, uint16_t dst_max)
 {
     if(src < src_min)
         return dst_min;
@@ -15,4 +15,3 @@ uint16_t map(uint16_t src, uint8_t src_min, uint8_t src_max,
 
     return dst_min + dst_delta*(src-src_min)/src_delta;
 }
-
