+++ robot/utils/map.h
@@ -3,7 +3,7 @@
 
 #include <stdint.h>
 
-uint16_t map(uint16_t src, uint8_t src_min, uint8_t src_max,
-        uint8_t dst_min, uint8_t dst_max);
+uint16_t map(uint16_t src, uint16_t src_min, uint16_t src_max,
+        uint16_t dst_min, uint16_t dst_max);
 
 #endif
